Zsh: comment out Weechat auto-start
This is part of an effort to make the dotfiles more minimal and more suited for
the main UUID 1000 user. We won't be launching the Weechat tmux session on
startup anymore by default, and it is now up to users to decide if they want
it. It is more logical to do it this way since Weechat itself requires some
configuration prior to enabling auto-start.